fix: builds on Ubuntu 24.10
Manages to patch things up just enough to get the whole thing to build on Ubuntu 24.10, on amd64.
The `./configure` command I used is quite similar to the macOS command in the README:
```
    ./configure --enable-tcg-interpreter --extra-ldflags=-g \
    --with-coroutine=gthread --enable-debug-tcg \
    --enable-debug --disable-werror --target-list="arm-softmmu" \
    --extra-cflags=-DDEBUG_CLKTREE --extra-cflags=-DDEBUG_STM32_RCC \
    --extra-cflags=-DDEBUG_STM32_UART --extra-cflags=-DSTM32_UART_NO_BAUD_DELAY \
    --extra-cflags=-DDEBUG_GIC
```

I cannot guarantee that everything works as expected - and there are some hacks I have to figure out (notably: what is with the hw/arm/stm32fxxxx.c files?) but it _does_ build and that's a start.